+++ index.html
@@ -4,6 +4,7 @@
   <meta charset="utf-8">
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
+  <link rel="stylesheet" href="vendor/leaflet/leaflet.css" crossorigin=""/>
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кексобукинг</title>
 </head>
@@ -238,5 +239,6 @@
 </body>
 
 <script src="./vendor/pristine/pristine.min.js"></script>
+<script src="./vendor/leaflet/leaflet.js" crossorigin=""></script>
 <script src="/js/main.js" type="module"></script>
 </html>
